requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
# crewai==0.1.26
python-dotenv==1.0.0

# Database
# psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.7.0

# AI/ML
sentence-transformers==2.2.2
spacy==3.7.2
scikit-learn==1.3.2
xgboost==2.0.3
transformers==4.35.2

# Document Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Task Queue
celery==5.3.4

# Utilities
pydantic==2.5.0
python-multipart==0.0.6
